added y overflow scroll in columns without showing scroll bar; make sure to only show up to 15 characters in task card; call api to get project data after moving a task to a new column
